requirements.txt
# Flask及核心扩展
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Moment==1.0.5
Flask-CKEditor==0.4.6
Flask-SocketIO==5.3.3
Flask-CORS==3.0.10

# 安全相关
Flask-Bcrypt==1.0.1
PyJWT==2.6.0
Flask-Security-Too==5.1.0
flask-reCaptcha==0.4.2

# 数据库和ORM
SQLAlchemy==2.0.4
alembic==1.9.4
pymysql==1.0.3

# 表单处理
WTForms==3.0.1
email-validator==1.3.1

# 文件处理
Pillow==9.4.0
python-magic==0.4.27
qrcode==7.4.2

# 工具
python-dotenv==1.0.0
python-dateutil==2.8.2
Faker==17.6.0
pendulum==2.1.2
bleach==6.0.0
markdown==3.4.1
pycountry==22.3.5

# API
marshmallow==3.19.0
Flask-RESTful==0.3.9
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# 测试
pytest==7.3.1
pytest-flask==1.2.0
coverage==7.2.2

# 生产环境
gunicorn==20.1.0
gevent==22.10.2
redis==4.5.1
celery==5.2.7
supervisor==4.2.5 